Feature/ml-105:  add 'clear dataset cache' option

+++ utils/huggingface/converters.py
@@ -11,10 +11,16 @@ from utils.helpers import ensure_dir, kili_print
 
 
 def kili_assets_to_hf_ner_dataset(
-    api_key: str, job: Dict, job_name: str, path_dataset: str, assets: List[Dict]
+    api_key: str,
+    job: Dict,
+    job_name: str,
+    path_dataset: str,
+    assets: List[Dict],
+    clear_dataset_cache: bool,
 ):
 
-    if os.path.exists(path_dataset):
+    if clear_dataset_cache and os.path.exists(path_dataset):
+        kili_print("Dataset cache for this project is being cleared.")
         os.remove(path_dataset)
 
     job_categories = list(job["content"]["categories"].keys())
@@ -26,57 +32,62 @@ def kili_assets_to_hf_ner_dataset(
 
     labels_to_ids = {label: i for i, label in enumerate(label_list)}
 
-    with open(ensure_dir(path_dataset), "w") as handler:
-        for asset in tqdm(assets):
-            response = requests.get(
-                asset["content"],
-                headers={
-                    "Authorization": f"X-API-Key: {api_key}",
-                },
-            )
-            text = response.text
-            annotations = asset["labels"][0]["jsonResponse"][job_name]["annotations"]
-            sentences = sent_tokenize(text)
-            offset = 0
-            for sentence_tokens in TreebankWordTokenizer().span_tokenize_sents(
-                sentences
-            ):
-                tokens = []
-                ner_tags = []
-                for start_without_offset, end_without_offset in sentence_tokens:
-                    start, end = (
-                        start_without_offset + offset,
-                        end_without_offset + offset,
-                    )
-                    token_annotations = [
-                        a
-                        for a in annotations
-                        if a["beginOffset"] <= start
-                        and a["beginOffset"] + len(a["content"]) >= end
-                    ]
-                    if len(token_annotations) > 0:
-                        category = token_annotations[0]["categories"][0]["name"]
-                        label = (
-                            "B-" + category
-                            if token_annotations[0]["beginOffset"] == start
-                            else "I-" + category
-                        )
-                    else:
-                        label = "O"
-                    tokens.append(text[start:end])
-                    ner_tags.append(labels_to_ids[label])
-                handler.write(
-                    json.dumps(
-                        {
-                            "tokens": tokens,
-                            "ner_tags": ner_tags,
-                        }
-                    )
-                    + "\n"
-                )
-                offset = offset + sentence_tokens[-1][1] + 1
+    if os.path.exists(path_dataset) and clear_dataset_cache:
+        os.remove(path_dataset)
+    if not os.path.exists(path_dataset):
+        with open(ensure_dir(path_dataset), "w") as handler:
+            for asset in tqdm(assets):
+                write_asset(api_key, job_name, labels_to_ids, handler, asset)
 
     return label_list
+
+
+def write_asset(api_key, job_name, labels_to_ids, handler, asset):
+    response = requests.get(
+        asset["content"],
+        headers={
+            "Authorization": f"X-API-Key: {api_key}",
+        },
+    )
+    text = response.text
+    annotations = asset["labels"][0]["jsonResponse"][job_name]["annotations"]
+    sentences = sent_tokenize(text)
+    offset = 0
+    for sentence_tokens in TreebankWordTokenizer().span_tokenize_sents(sentences):
+        tokens = []
+        ner_tags = []
+        for start_without_offset, end_without_offset in sentence_tokens:
+            start, end = (
+                start_without_offset + offset,
+                end_without_offset + offset,
+            )
+            token_annotations = [
+                a
+                for a in annotations
+                if a["beginOffset"] <= start
+                and a["beginOffset"] + len(a["content"]) >= end
+            ]
+            if len(token_annotations) > 0:
+                category = token_annotations[0]["categories"][0]["name"]
+                label = (
+                    "B-" + category
+                    if token_annotations[0]["beginOffset"] == start
+                    else "I-" + category
+                )
+            else:
+                label = "O"
+            tokens.append(text[start:end])
+            ner_tags.append(labels_to_ids[label])
+        handler.write(
+            json.dumps(
+                {
+                    "tokens": tokens,
+                    "ner_tags": ner_tags,
+                }
+            )
+            + "\n"
+        )
+        offset = offset + sentence_tokens[-1][1] + 1
 
 
 def predicted_tokens_to_kili_annotations(
@@ -127,7 +138,9 @@ def predicted_tokens_to_kili_annotations(
                 len(kili_annotations)
                 and ann["categories"][0]["name"]
                 == kili_annotations[-1]["categories"][0]["name"]
-                and (ann["beginOffset"] == kili_annotations[-1]["endOffset"] or is_i_tag)
+                and (
+                    ann["beginOffset"] == kili_annotations[-1]["endOffset"] or is_i_tag
+                )
             ):
                 # merge with previous if same category and contiguous offset and onset:
                 kili_annotations[-1]["endOffset"] = ann["endOffset"]
